Split request and response detector tests
